requirements.txt
pandas==1.5.3
streamlit==1.20.0
plotly.express==0.4.1
firebase-admin==6.1.0
google-cloud-firestore==2.10.0
Faker